// tdg-management-api-backend/src/sprints/types/request.type.ts
export enum SprintSortBy {
  START_DATE_ASC = 'startDateAsc',
  START_DATE_DESC = 'startDateDesc',
  END_DATE_ASC = 'endDateAsc',
  END_DATE_DESC = 'endDateDesc',
  ESTIMATED_START_DATE_ASC = 'estimatedStartDateAsc',
  ESTIMATED_START_DATE_DESC = 'estimatedStartDateDesc',
  ESTIMATED_END_DATE_ASC = 'estimatedEndDateAsc',
  ESTIMATED_END_DATE_DESC = 'estimatedEndDateDesc',
  CREATED_AT_ASC = 'createdAtAsc',
  CREATED_AT_DESC = 'createdAtDesc',
}